Separate views into right files
